TEST: Move mock to correct test
